moved the vote commands from Program to VoteManager
